feat: component information place
